requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-jose==3.5.0
redis==6.2.0

pydantic==2.11.7
pydantic-core==2.33.2
typing-extensions==4.14.0
typing-inspection==0.4.1

anyio==4.9.0
starlette==0.46.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib==1.7.4
